TST: test_to_time (#32285)
